refactored cv3_family operation into multiple operations
